float scale

+++ main_mobile.go
@@ -21,7 +21,7 @@ func IsRunning() bool {
 	return running
 }
 
-func Start(scale int) error {
+func Start(scale float64) error {
 	running = true
 	game, err := ino.NewGame()
 	if err != nil {
